improve types, flatten send request
